Removed support for SDK 0.x
